moved form validation to utilities folder
